feat: refactor New Arrivals page using async/await and add NewArrivalsClient component for improved UX
